arrow key movement
paddle boundaries changed

https://en.wikipedia.org/wiki/Pong

Another feature was that the in-game paddles were unable to reach the top of the screen. This was caused by a simple circuit that had an inherent defect. Instead of dedicating time to fixing the defect, Alcorn decided it gave the game more difficulty and helped limit the time the game could be played; he imagined two skilled players being able to play forever otherwise.

http://www.ign.com/articles/2008/03/11/al-alcorn-interview